+++ crates/entrypoint/src/cli/command/azure_devops/license_entitlement/user/mod.rs
@@ -2,9 +2,11 @@ pub mod list;
 pub mod revoke;
 pub mod show;
 pub mod update;
+mod azure_devops_license_entitlement_user_matcher;
 
 pub use list::AzureDevOpsLicenseEntitlementUserListArgs;
 pub use revoke::AzureDevOpsLicenseEntitlementUserRevokeArgs;
 pub use show::AzureDevOpsLicenseEntitlementUserShowArgs;
 pub use update::AzureDevOpsLicenseEntitlementUserUpdateArgs;
 pub use update::AzureDevOpsLicenseEntitlementUserUpdateTuiArgs;
+pub use azure_devops_license_entitlement_user_matcher::*;