+++ CHANGELOG.md
@@ -9,12 +9,14 @@
     - `ct az devops license-entitlement user list`
     - `ct az devops license-entitlement user update`
     - `ct az devops license-entitlement user update-tui`
-    - `ct az devops license-entitlement user show --id {}`
+    - `ct az devops license-entitlement user show --user-devops-id {}`
+    - `ct az devops license-entitlement user show --user-email {}`
     - `ct az devops license-entitlement group list`
 - Rename `AzureDevOpsLicenseKind` to `AzureDevOpsLicenseType`, add stronger typing
 - Fix Microsoft graph helper failing to deserialize error bodies
 - Add cache key support to Microsoft graph helper
-- Add `ct az devops license-entitlement user revoke --user-id {}`
+- Add `ct az devops license-entitlement user revoke --user-devops-id {}`
+- Add `ct az devops license-entitlement user revoke --user-email {}`
 - Add `ct az ad group member add --group-id {} --member-id {}` and supporting aspects
 - Add `ct az ad group member remove --group-id {} --member-id {}` and supporting aspects
 - Change `--log-file` to append instead of truncate